[css-grid] Refactoring of track sizing state machine
Some of the state transition validation was not useful, since the track
sizing algorithm never goes through that. It was intended for repeat the
first iteration of the row track sizing in case of an indefinite height
grid container, but since intrinsic height computation does not touch
the state, such repeated transition is not needed.

Additionally, is not correct that a function which purpose is to check
out the validity of a state modifies such state.

This is just a refactoring so no additional tests are needed, neither
there is an associated bug.

Review-Url: https://codereview.chromium.org/2186003002
Cr-Original-Commit-Position: refs/heads/master@{#408140}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aa03236b385ab74d7e54e67ebdef6d01b01c9c8b