Dont bail out of preload scanning if a appcache manifest is found
The preload scanner bails if it sees an appcache manifest in the markup.
This behavior isn't really correct, as we already wait for appcache to
be initialized before sending out any requests. Also, there could be
resources that still need to be loaded over the network.

BUG=630717

Review-Url: https://codereview.chromium.org/2187193002
Cr-Original-Commit-Position: refs/heads/master@{#408304}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 91146fe960885eae005f3ac14f29dacc93a98dac