Implement resize option for createImageBitmap from ImageData
In the case of ImageData, we use two different code path for the
following two cases:
1). ImageBitmapOptions.premultiplyAlpha = true
2). ImageBitmapOptions.premultiplyAlpha = false

In the first case, we create a SkSurface and draw the original SkImage
onto it for resizing. In the second case, we call scalePixels on the
original SkImage.

BUG=627855

Review-Url: https://codereview.chromium.org/2183323003
Cr-Original-Commit-Position: refs/heads/master@{#408402}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5c5fa614b44d8133ab31280b67fe33638621a192