Change the expected results on win10 bots
Right now fast/images/paint-subrect-grid.html fails on win10. By looking
at the expected results directory, I found that win7 doesn't have
expected.png for this test, which means it refers to the expected.png
under win/ directory. By looking at the stdio of win7 bots, it appears
that this test always pass at the first run. That means the expected.png
is good for win7, but not good for win10.

So in this CL, we copy the expected.png from win/ to win7/. And replace
the expected.png under win/ by the actual results generated from the
win10 bots. So that next time when win10 bots runs, it will refer to
the actual result that itself generated, which should pass.
We should closely monitor win7 and win10 bots to make sure that
they are green after this change.

TBR=schenney@chromium.org
BUG=632000

Review-Url: https://codereview.chromium.org/2188893003
Cr-Original-Commit-Position: refs/heads/master@{#408450}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fe248e2d21e6c1852584c47e628e6b8664cc7faf