[DevTools] implement isSpaceOrNewline for the STL string wrapper.
We now use std::isspace, which is not equivalent to the WTF
implementation. But it is consistent with the STL version of
String16::stripWhiteSpace.

With this, the inspector protocol parser test succeeds in the
V8 standalone build that uses the STL string wrapper.

R=dgozman@chromium.org
BUG=580337

Review-Url: https://codereview.chromium.org/2188203002
Cr-Original-Commit-Position: refs/heads/master@{#408591}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fc6d7b372defe05402d8d035608a3af5f2e2f9f2