[Fetch API] Introduce BytesConsumer
This CL introduces BytesConsumer interface which is expected to replace most of
WebDataConsumerHandle subclasses. Unlike WebDataConsumerHandle, BytesConsumer
is bound to a thread and an on-heap interface.

This CL also implements a bridge class from FetchDataConsumerHandle to
BytesConsumer. As it is not yet used, this CL doesn't change the actual
behavior.

BUG=610195

Review-Url: https://codereview.chromium.org/2046203003
Cr-Original-Commit-Position: refs/heads/master@{#408592}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e14ef20c8ff5251390825f79e83f7db23e28dbf0