Skip dirtying lines when attaching children prior to attaching oneself
If a parent is about to be (re)attached then there is no point in keeping its
dirty lines up to date as we attach children. The parent will get a full layout
anyway and we end up having to dirty the lineboxes for every inlines we add - which
can get expensive if there are a lot of inlines.

BUG=345972

Review-Url: https://codereview.chromium.org/1550513003
Cr-Original-Commit-Position: refs/heads/master@{#408601}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dcb4562e8f888588dfe686b1b9d864e39b3dc5cb