In rietveld.filter_latest_jobs, handle jobs with no build number.
Currently, in some circumstances, filter_try_jobs may throw an error depending on the state of the current try jobs.

This CL changes just this method so that it doesn't throw a KeyError if there are jobs with no build number passed in.

Review-Url: https://codereview.chromium.org/2182413004
Cr-Original-Commit-Position: refs/heads/master@{#408874}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e5e73d3df3d7d0a42f28040ab0b6fe022c1179c5