Clarify fullscreenElement vs. currentFullScreenElement distinction
No observable changes are intended.

This may make DocumentFullscreen::fullscreenElement and
::currentFullScreenElement slightly slower, but if that turns out to be
a problem a reasonable fix would be to inline the corresponding getters
in Fullscreen.h.

The Screen in m_fullScreenElementStack is lowercased, as it should be
and it shows all the places where it is used internally.

BUG=402421
R=eae@chromium.org

Review-Url: https://codereview.chromium.org/2199783002
Cr-Original-Commit-Position: refs/heads/master@{#408993}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 63ec834253d3a3da461f96ec5f659b524bf8ad58