Fix crash with detached HTMLDocumentParser
When a detached HTMLDocumentParser gets finalized the destructor calls
HTMLDocumentParser::stopBackgroundParser which tries to tell the
scheduler, unfortunatly document() is null and it crashes.

We can avoid this by checking if m_haveBackgroundParser is true before
trying to tell the scheduler.  This fixes a bug where closing a
HTMLDocumentParser without a background parser does the wrong thing,
and it also fixes the crash since if m_haveBackgroundParser is true,
DocumentParser::detach hasn't been called yet and document() is not
null.  See HTMLDocumentParser::detech for details.

BUG=633534

Review-Url: https://codereview.chromium.org/2198273003
Cr-Original-Commit-Position: refs/heads/master@{#409222}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1dc3b9489cbcfff1e3258c17145e04e0d1d87f9c