Add a new class AcceleratedStaticBitmap : StaticBitmapImage
At this moment, the StaticBitmapImage holds RefPtr to a SkImage, or a
WebExternalTextureMailbox. So there are in total of three cases:
1). The SkImage is not texture backed.
2). The SkImage is GPU texture backed.
3). The SkImage is null, but has a valid mailbox.

It makes more sense to break this class into two classes. So this CL
adds a new class called AcceleratedStaticBitmap, inherited from the
StaticBitmapImage class, is in charge of case 2 and 3. Case 1 is handled
by the StaticBitmapImage class.

We can use the bots (including the GPU_Optional bots) to make sure this
change doesn't break anything.

BUG=632722

Review-Url: https://codereview.chromium.org/2196963002
Cr-Original-Commit-Position: refs/heads/master@{#409238}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 23b4e88d6090ae9f2b2fa9aa062152900c09f7e9