Change consumeUrl to return a StringView instead of a String
Since consumeUrl is always returning a string from a CSSParserToken or
a null string, it makes more sense to return a StringView instead of a
String. For cases where we convert the returned value to a AtomicString,
this saves us an intermediate conversion to a String object.

Review-Url: https://codereview.chromium.org/2202763004
Cr-Original-Commit-Position: refs/heads/master@{#409259}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aeee7861e2f53195e580d36b2b6d05854ca0dfe5