Don't let rounding prematurely influence document size when printing
Don't round away subpixel values when they can still make a difference to the
size of the document. We convert back and forth between different DPIs so we
can lose a pixel or two along the way if we don't retain subpixel values to
the end.

BUG=467579

Review-Url: https://codereview.chromium.org/2116283002
Cr-Original-Commit-Position: refs/heads/master@{#409469}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 53c779cbb1a137fd5781422c67f36c3354382bc8