Remove String::charactersWithNullTermination.
All of the callers except LocaleWin just want to convert to UChar and don't care
about the null termination. We can use ensure16Bit for that which has the added
benefit that strings which were already 16bit don't require a malloc and copy.

For LocaleWin we can convert it to using an on stack buffer instead which also saves
the malloc.

Doing this also exposed a bug in String::ensure16Bit where an 8bit empty string
would not get converted to a 16bit empty string, so doing:

String string(StringImpl::empty())
string.ensure16Bit();
string.characters16(); // Fail assert for !is8Bit().

This is fixed in this patch and unit tests are added.

Review-Url: https://codereview.chromium.org/2200313003
Cr-Original-Commit-Position: refs/heads/master@{#409519}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c57e2b0d8874358c3ee32cc2126212a0a9a8e2de