Pass template files explicitly to Robolectric template processor.
Previously, the template processor searched through a directory you
passed to it and looked for all .vm files. Changing it so you have
to explicitly pass the lits of .vm files to the template processor
script. Verified this will fix a problem with the WebRTC bots not
being able to find the template files.

Review-Url: https://codereview.chromium.org/2200413004
Cr-Original-Commit-Position: refs/heads/master@{#409687}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c46e862757edc04c06b1bd88724d15a5807b84d1