Modify LayoutTests to prep for compositing solid background scollers
Tests of scrolling behavior should allow some visual mechanism for
detecting the scroll position. Some tests have been updated to make this
so.

Other tests require that a scrolling layer not be composited. Upcoming
changes will cause solid background color scrolling content to be
composited on all displays, so such tests must be modified slightly to
prevent the scrolling layer from promotion (by making the background
partially transparent).

And one test is explicitly testing the preferCompositingToLCDText flag,
so we should rename it and adjust the background so that the test
continues to test what it claims.

R=flackr
BUG=381840

Review-Url: https://codereview.chromium.org/2196363002
Cr-Original-Commit-Position: refs/heads/master@{#409791}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a0f3f2e7a62356e4c1a44b769e46be984366ecad