[css-flexbox] Simplify the "child intrinsic height" logic
The logic in childIntrinsicHeight/childIntrinsicWidth was very
hard to understand. This patch makes them work on the logical
width/height instead, which is much simpler.

This also removes a TODO mentioning indefinite percentages, because
the spec is clear that only actual auto values should be stretched.

R=eae@chromium.org,dgrogan@chromium.org

Review-Url: https://codereview.chromium.org/2216133003
Cr-Original-Commit-Position: refs/heads/master@{#410348}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 73e356aec7507c9e27dcbd9728511e9775d03a44