Fix properties of got/lostpointercapture
Make the attributes of got/lostpointercapture the same
as the event that caused them to be fired. In the case
of the delayed processing that will be the very next
pointerevent. In the case of implicit release that
will be pointerup/cancel right before got/lostpointercapture
event.

BUG=632766

Review-Url: https://codereview.chromium.org/2199263005
Cr-Original-Commit-Position: refs/heads/master@{#410409}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3bd09006149797436511e02f4196ce54a91cd72e