Improve canBeAnchorNode() for PositionInFlatTree constructor
This patch changes |canBeAnchorNode()| for |PositionInFlatTree| constructor
to check whether specified node can be participate to flat tree or not to
improve debug experience.

This patch is a preparation of http://crrev.com/2221783003, which attempt to
call |toPositionInFlatTree()| with a position which uses |ShadowRoot| as
anchor node.

BUG=139552, 632333
TEST=n/a; no behavior changes

Review-Url: https://codereview.chromium.org/2229703002
Cr-Original-Commit-Position: refs/heads/master@{#410629}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d0bf55edc59dd1399c5892f1afb2643ea978bddb