Set objects to need full paint invalidation after clearing their invalidation rects.
This is necessary in the world of rtrees driven by visual rects, because it's
now required that visual rects be correct on every frame. Previously we did not
need this to happen, because we just needed to know if an invalidation changed
the visual rect.

BUG=634959

Review-Url: https://codereview.chromium.org/2220333004
Cr-Original-Commit-Position: refs/heads/master@{#410712}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2234389b8ef6a9204d9a37bb17810ab9fdd70441