CacheStorage: Expand cache.keys() method.
Current implementation of the method is always return all of keys even if it
takes optional parameters. According to ServiceWorkerSpec[1], the method should
take a request object and cache query options[2] as optional parameters.
This CL is expanding the method as per the spec.

[1] http://slightlyoff.github.io/ServiceWorker/spec/service_worker/#cache-keys
[2] Currenly, only supports ignoreSearch option like matchAll(), delete()

BUG=520784

Review-Url: https://codereview.chromium.org/1719103002
Cr-Original-Commit-Position: refs/heads/master@{#410868}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e6a4575cd18e48c18e7d5f5183e2fc17e537b725