Refactoring: Remove dead code in DeferredImageDecoder::createFrameAtIndex
DeferredImageDecoder::createFrameAtIndex is used only at
ImageSource::createFrameAtIndex, and this is indirectly called by
BitmapImage::frameAtIndex, which checks that the index is less than the
number of frames. Thus, DeferredImageDecoder::createFrameAtIndex can
assume that the given |index| is less than the number of the frames.

This CL removes the dead code to improve code health.

BUG=n/a
TEST=webkit_unit_tests

Review-Url: https://codereview.chromium.org/2225903004
Cr-Original-Commit-Position: refs/heads/master@{#410961}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9a1a7dc0502f5d16b53ce50a2fd509b0ea5cf125