CSS: animated text-decoration-line overrides inline style
The text-decoration-line of an element can add to the inherited
text-decoration-line. However, StyleAdjuster::adjustComputedStyle
is called both before and after animations are applied, and the
text-decoration-line from the inline style (or SVG attribute)
was being combined with the text-decoration-line from the CSS
or Web Animation.

BUG=620037

Review-Url: https://codereview.chromium.org/2227003002
Cr-Original-Commit-Position: refs/heads/master@{#410966}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5243763790094312f18657ffa69c21ba4f613566