Handle resize option when  input is an SVG image
When the input of createImageBitmap is a SVG image and it has no
intrinsic size or zero intrinsic size, it is required that either the
resize option is given or the cropRect is given.

This CL makes change such that it will throw a InvalidState error when
none of them is specified.

BUG=627855

Review-Url: https://codereview.chromium.org/2210623002
Cr-Original-Commit-Position: refs/heads/master@{#411022}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c80c0c565168956900daff3b37245a76e86f3fcd