Reorganize test expectations to consolidate paint team tests
The aim is to consolidate tests owned by the paint team to make future change tracking easier. It is a goal of the team to track layout team regressions, additions and fixes.

This patch should have no net effect on test expectations. It is a line permutation only.

R=chrishtr@chromium.org
BUG=636475

Review-Url: https://codereview.chromium.org/2233053002
Cr-Original-Commit-Position: refs/heads/master@{#411123}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e01bd4dbaa7ea4468e69dfa2447234cf0cecb223