Skip redundant fonts returned by FontFallbackIterator
HarfBuzzShaper asks FontFallbackIterator::next for a font by calling
next() each time it is done shaping with this font and still has
non-rendered glyphs. Depending on what the font-stack, user-preference
font, and fallback fonts returned by the system are, the shaper may try
to shape with the same font several times, and still end up not finding a
result before the LastResort font.

Since HarfBuzzFace objects are cached internally to that class and
unified using Skia unique SkTypeface IDs, we can filter for SkTypeFace
uniqueIds earlier in FontFallbackIterator and skip redundant shaping.

Examples would be:
Empty font-family CSS, but CSS code seems to set the default serif font
before we reach shaping, then we try with that, and afterwards attempt
to shape with the users default font again, which is the same, Times New
Roman in both cases. Similarly, a font-family: that has Times or Times
New Roman which we alias to the same font, etc.

BUG=636347
TEST=Manually tested this by adding debug code to HarfBuzzShaper and
detecting incoming duplicates.

Review-Url: https://codereview.chromium.org/2230233002
Cr-Original-Commit-Position: refs/heads/master@{#411164}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5fdb26babe2d64e6f09175edffd5efe4abf220d8