Update create_components_summary.py to allow other tree names in bower.json
It's nice to be able to test a particular branch of a Polymer repo, by updating
bower.json locally and running reproduce.sh, but doing so causes
create_components_summary.py to fail loudly due to missing properties in the
.bower.json.

We should find the more general tree-ish name instead of the "tag" property.
For release tags (what our committed bower.json uses) this is the same.

R=dbeam@chromium.org

Review-Url: https://codereview.chromium.org/2225513002
Cr-Original-Commit-Position: refs/heads/master@{#411176}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b561abad688e9fca08fa7663e8f6e921f8b61e7b